COM Interface Name Conflicts
Some COM Interfaces start with an underscore which was automatically removed, because Crystal does not allow the first character of a struct name to be an underscore, but it appears that it was added originally for name conflicts. The affected COM Interfaces have an extra prepended 'I' to its name to resolve conflicts.