Added documentation to macros file
Added documentation to the methods `hiword` and `loword`.
Added pointer versions of the `hiword` and `loword` methods.
Added documentation to the `pwstr` macro.
Added specs tests for `hiword` and `loword` methods and the `pwstr` macro.

Found a few issues that were corrected as documented below.
Corrected the input parameters from `Int64` to `Int32`, or `UInt64` to `UInt32`, for one of the hiword and loword methods.
Corrected conversions to be `to_u32` instead of `to_u64`.